refactor(backend): extract ConversationWriter and ThreadReplay
Completes the split. `persistConversation` and `appendConversationMessage` move
to `conversation-writer.ts`, which takes the roster as a hard constructor
dependency for the agent lookup and `ensureThreadProjection`;
`rebuildThreadProjection` moves to `thread-replay.ts`, which takes the
conversation queries because the replay ends by reading the thread back.

`openbot-database.ts` is now a 313-line facade over eight controllers plus the
core: the class name, instance identity, constructor signature and public
surface are unchanged, which is what `withDatabaseTransaction` relies on when it
keys its `WeakMap` on the instance.

`persistConversationAndMailbox` stays on the facade — it is the one place two
controllers must share a single unconditional transaction, and its `BEGIN
IMMEDIATE` is what makes both nested dispatches join it instead of committing
piecemeal.
