+++ src/backend/openbot-database.ts
@@ -10,21 +10,11 @@ import type {
   ConversationSnapshot,
   HostedSiteConversationEventStatus,
 } from "@openbot/contracts/ipc";
-import { isAgentProvider, isConversationMessage, providerForLegacyModel } from "@openbot/contracts/ipc";
-import { type DynamicRecord, isDynamicRecord, isNumber, isString } from "@openbot/contracts/runtime-values";
+import { providerForLegacyModel } from "@openbot/contracts/ipc";
 import { AgentRoster } from "./database/agent-roster";
 import { ConversationQueries } from "./database/conversation-queries";
-import { DatabaseCore, deleteOrphanReceipts, type OrchestrationEventInput } from "./database/database-core";
-import {
-  databaseRow,
-  databaseRows,
-  decodeThreadAgentRow,
-  objectValue,
-  optionalStringColumn,
-  requiredEventRow,
-  requiredNumberColumn,
-  requiredStringColumn,
-} from "./database/database-rows";
+import { ConversationWriter } from "./database/conversation-writer";
+import { DatabaseCore, type OrchestrationEventInput } from "./database/database-core";
 import {
   type ActiveHostedSiteConversationEvent,
   HostedSiteEventLog,
@@ -32,6 +22,7 @@ import {
 } from "./database/hosted-site-event-log";
 import { MailboxProjection, type MailboxProjectionState } from "./database/mailbox-projection";
 import { type ProviderSession, ProviderSessions } from "./database/provider-sessions";
+import { ThreadReplay } from "./database/thread-replay";
 import { type StoredThreadSummary, ThreadSummaries } from "./database/thread-summaries";
 
 // Declared in this module before the split and part of the frozen public surface, so it stays
@@ -55,6 +46,8 @@ export class OpenBotDatabase {
   readonly #core: DatabaseCore;
   readonly #conversations: ConversationQueries;
   readonly #roster: AgentRoster;
+  readonly #conversationWrites: ConversationWriter;
+  readonly #replay: ThreadReplay;
   readonly #hostedSiteEvents: HostedSiteEventLog;
   readonly #mailbox: MailboxProjection;
   readonly #sessions: ProviderSessions;
@@ -64,6 +57,8 @@ export class OpenBotDatabase {
     this.#core = new DatabaseCore({ userDataPath });
     this.#conversations = new ConversationQueries({ core: this.#core });
     this.#roster = new AgentRoster({ core: this.#core });
+    this.#conversationWrites = new ConversationWriter({ core: this.#core, roster: this.#roster });
+    this.#replay = new ThreadReplay({ core: this.#core, conversations: this.#conversations });
     this.#hostedSiteEvents = new HostedSiteEventLog({ core: this.#core });
     this.#mailbox = new MailboxProjection({ core: this.#core });
     this.#sessions = new ProviderSessions({ core: this.#core });
@@ -198,134 +193,7 @@ export class OpenBotDatabase {
     payload: unknown = {},
     commandId = `conversation:${eventType}:${randomUUID()}`,
   ): ConversationSnapshot {
-    if (!snapshot.threadId) return structuredClone(snapshot);
-    const threadId = snapshot.threadId;
-    const recovery = conversationRecoveryState(this.connection, threadId, snapshot.activeTurnId);
-    const result = this.dispatch(
-      commandId,
-      [
-        {
-          aggregateType: "thread",
-          aggregateId: threadId,
-          eventType,
-          payload: { detail: payload, snapshot, recovery },
-        },
-      ],
-      (db, sequences) => {
-        const sequence = sequences[0] ?? snapshot.revision;
-        const agent = this.#roster.listAgents().find((candidate) => candidate.id === snapshot.botId);
-        if (!agent) throw new Error(`Unknown agent for conversation: ${snapshot.botId}`);
-        this.#roster.ensureThreadProjection(db, agent, sequence);
-        db.prepare(
-          `UPDATE projection_threads
-           SET active_turn_id = ?, updated_at = ?, last_event_sequence = ? WHERE thread_id = ?`,
-        ).run(snapshot.activeTurnId, new Date().toISOString(), sequence, snapshot.threadId);
-        const messageIds = new Set(snapshot.messages.map((message) => message.id));
-        const staleMessageIds = databaseRows(
-          db.prepare("SELECT message_id FROM projection_thread_messages WHERE thread_id = ?").all(threadId),
-        )
-          .map((row) => requiredStringColumn(row, "message_id"))
-          .filter((messageId) => !messageIds.has(messageId));
-        const deleteMessage = db.prepare(
-          "DELETE FROM projection_thread_messages WHERE thread_id = ? AND message_id = ?",
-        );
-        const deleteAttachments = db.prepare(
-          "DELETE FROM projection_attachments WHERE owner_kind = 'thread-message' AND owner_id = ?",
-        );
-        for (const messageId of staleMessageIds) {
-          deleteAttachments.run(`${threadId}:${messageId}`);
-          deleteMessage.run(threadId, messageId);
-        }
-        const upsert = db.prepare(`
-          INSERT INTO projection_thread_messages (
-            thread_id, message_id, turn_id, author, status, item_type, created_at,
-            ordinal, message_json, last_event_sequence
-          ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-          ON CONFLICT(thread_id, message_id) DO UPDATE SET
-            turn_id = excluded.turn_id,
-            author = excluded.author,
-            status = excluded.status,
-            item_type = excluded.item_type,
-            created_at = excluded.created_at,
-            ordinal = excluded.ordinal,
-            message_json = excluded.message_json,
-            last_event_sequence = excluded.last_event_sequence
-        `);
-        snapshot.messages.forEach((message, ordinal) => {
-          upsert.run(
-            snapshot.threadId,
-            message.id,
-            message.turnId ?? null,
-            message.author,
-            message.status,
-            message.itemType ?? null,
-            message.createdAt,
-            ordinal,
-            JSON.stringify(message),
-            sequence,
-          );
-          for (const attachment of message.attachments ?? []) {
-            db.prepare(`
-              INSERT OR REPLACE INTO projection_attachments
-                (attachment_id, owner_kind, owner_id, name, path, metadata_json, created_at, last_event_sequence)
-              VALUES (?, 'thread-message', ?, ?, '', ?, ?, ?)
-            `).run(
-              `${snapshot.threadId}:${message.id}:${attachment.id}`,
-              `${snapshot.threadId}:${message.id}`,
-              attachment.name,
-              JSON.stringify(attachment),
-              message.createdAt,
-              sequence,
-            );
-          }
-        });
-        db.prepare(`
-          INSERT INTO projection_thread_activities
-            (activity_id, thread_id, turn_id, activity_type, payload_json, created_at, last_event_sequence)
-          VALUES (?, ?, ?, ?, ?, ?, ?)
-        `).run(
-          randomUUID(),
-          snapshot.threadId,
-          snapshot.activeTurnId,
-          eventType,
-          JSON.stringify(payload),
-          new Date().toISOString(),
-          sequence,
-        );
-        if (snapshot.activeTurnId) {
-          db.prepare(`
-            INSERT INTO projection_turns
-              (turn_id, thread_id, provider_session_id, status, started_at, completed_at, last_event_sequence)
-            VALUES (?, ?, (
-              SELECT id FROM projection_provider_sessions
-              WHERE thread_id = ? AND state = 'active' ORDER BY created_at DESC LIMIT 1
-            ), 'running', ?, NULL, ?)
-            ON CONFLICT(turn_id) DO UPDATE SET status = 'running', last_event_sequence = excluded.last_event_sequence
-          `).run(snapshot.activeTurnId, snapshot.threadId, snapshot.threadId, new Date().toISOString(), sequence);
-        }
-        if (
-          ["turn.completed", "turn.reconciled-after-restart", "turn.interrupted-by-restart"].includes(eventType) &&
-          isDynamicRecord(payload) &&
-          "turnId" in payload &&
-          isString(payload.turnId)
-        ) {
-          const status =
-            "status" in payload && isString(payload.status)
-              ? payload.status
-              : eventType === "turn.interrupted-by-restart"
-                ? "interrupted"
-                : "completed";
-          db.prepare(
-            `UPDATE projection_turns
-             SET status = ?, completed_at = ?, last_event_sequence = ?
-             WHERE thread_id = ? AND turn_id = ?`,
-          ).run(status, new Date().toISOString(), sequence, snapshot.threadId, payload.turnId);
-        }
-        pruneConversationSnapshots(db, threadId, sequence);
-        return { revision: sequence };
-      },
-    );
-    return { ...structuredClone(snapshot), revision: result.revision };
+    return this.#conversationWrites.persistConversation(snapshot, eventType, payload, commandId);
   }
 
   appendConversationMessage(input: {
@@ -337,89 +205,7 @@ export class OpenBotDatabase {
     detail?: unknown;
     commandId?: string;
   }): number {
-    const result = this.dispatch(
-      input.commandId ?? `conversation:${input.eventType}:${randomUUID()}`,
-      [
-        {
-          aggregateType: "thread",
-          aggregateId: input.threadId,
-          eventType: input.eventType,
-          occurredAt: input.message.createdAt,
-          payload: {
-            detail: input.detail ?? {},
-            appendedMessage: input.message,
-            activeTurnId: input.activeTurnId,
-          },
-        },
-      ],
-      (db, sequences) => {
-        const sequence = sequences[0] ?? 0;
-        const agent = this.#roster.listAgents().find((candidate) => candidate.id === input.botId);
-        if (!agent || agent.threadId !== input.threadId) {
-          throw new Error(`Unknown agent thread for conversation append: ${input.botId}`);
-        }
-        this.#roster.ensureThreadProjection(db, agent, sequence);
-        const ordinalRow = databaseRow(
-          db
-            .prepare(
-              `SELECT COALESCE(MAX(ordinal), -1) + 1 AS ordinal
-               FROM projection_thread_messages WHERE thread_id = ?`,
-            )
-            .get(input.threadId),
-        );
-        const ordinal = ordinalRow ? requiredNumberColumn(ordinalRow, "ordinal") : 0;
-        db.prepare(
-          `INSERT INTO projection_thread_messages (
-             thread_id, message_id, turn_id, author, status, item_type, created_at,
-             ordinal, message_json, last_event_sequence
-           ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
-        ).run(
-          input.threadId,
-          input.message.id,
-          input.message.turnId ?? null,
-          input.message.author,
-          input.message.status,
-          input.message.itemType ?? null,
-          input.message.createdAt,
-          ordinal,
-          JSON.stringify(input.message),
-          sequence,
-        );
-        for (const attachment of input.message.attachments ?? []) {
-          db.prepare(
-            `INSERT INTO projection_attachments
-               (attachment_id, owner_kind, owner_id, name, path, metadata_json, created_at, last_event_sequence)
-             VALUES (?, 'thread-message', ?, ?, '', ?, ?, ?)`,
-          ).run(
-            `${input.threadId}:${input.message.id}:${attachment.id}`,
-            `${input.threadId}:${input.message.id}`,
-            attachment.name,
-            JSON.stringify(attachment),
-            input.message.createdAt,
-            sequence,
-          );
-        }
-        db.prepare(
-          `UPDATE projection_threads
-           SET active_turn_id = ?, updated_at = ?, last_event_sequence = ? WHERE thread_id = ?`,
-        ).run(input.activeTurnId, input.message.createdAt, sequence, input.threadId);
-        db.prepare(
-          `INSERT INTO projection_thread_activities
-             (activity_id, thread_id, turn_id, activity_type, payload_json, created_at, last_event_sequence)
-           VALUES (?, ?, ?, ?, ?, ?, ?)`,
-        ).run(
-          randomUUID(),
-          input.threadId,
-          input.message.turnId ?? input.activeTurnId,
-          input.eventType,
-          JSON.stringify(input.detail ?? {}),
-          input.message.createdAt,
-          sequence,
-        );
-        return { revision: sequence };
-      },
-    );
-    return result.revision;
+    return this.#conversationWrites.appendConversationMessage(input);
   }
 
   persistConversationAndMailbox(
@@ -433,7 +219,7 @@ export class OpenBotDatabase {
     db.exec("BEGIN IMMEDIATE");
     try {
       this.replaceMailboxState(`mailbox:${mailboxEventType}:${randomUUID()}`, mailboxState, mailboxEventType);
-      const persisted = this.persistConversation(
+      const persisted = this.#conversationWrites.persistConversation(
         snapshot,
         eventType,
         payload,
@@ -488,225 +274,7 @@ export class OpenBotDatabase {
   }
 
   rebuildThreadProjection(threadId: string): ConversationSnapshot {
-    const db = this.connection;
-    const events = databaseRows(
-      db
-        .prepare(
-          `SELECT sequence, event_type, occurred_at, payload_json
-           FROM orchestration_events
-           WHERE aggregate_type = 'thread' AND aggregate_id = ? ORDER BY sequence`,
-        )
-        .all(threadId),
-    ).map(requiredEventRow);
-    const thread = decodeThreadAgentRow(
-      db.prepare("SELECT agent_id FROM projection_threads WHERE thread_id = ?").get(threadId),
-    );
-    if (!thread) throw new Error(`Unknown OpenBot thread: ${threadId}`);
-
-    let latest: ConversationSnapshot | null = null;
-    let latestSequence = 0;
-    const sessions = new Map<string, ProviderSession & { sequence: number }>();
-    const summaries: Array<StoredThreadSummary & { sequence: number }> = [];
-    const turnSessions = new Map<string, string | null>();
-    for (const event of events) {
-      const payload = JSON.parse(event.payload_json);
-      const record = objectValue(payload);
-      if (event.event_type === "provider-session.bound") {
-        const session = providerSessionValue(record);
-        if (session) {
-          for (const current of sessions.values()) current.state = "inactive";
-          sessions.set(session.id, { ...session, sequence: event.sequence });
-        }
-      } else if (event.event_type === "provider-session.deactivated") {
-        const ids = Array.isArray(record?.sessionIds) ? record.sessionIds : [];
-        for (const id of ids) {
-          if (isString(id)) {
-            const session = sessions.get(id);
-            if (session) session.state = "inactive";
-          }
-        }
-      } else if (event.event_type === "provider-session.config-updated") {
-        const session = isString(record?.sessionId) ? sessions.get(record.sessionId) : null;
-        if (session) {
-          if (isString(record?.model)) session.model = record.model;
-          if (isString(record?.effort)) session.effort = record.effort;
-          session.sequence = event.sequence;
-        }
-      } else if (event.event_type === "thread.summary-created") {
-        const summary = summaryValue(record);
-        if (summary) summaries.push({ ...summary, sequence: event.sequence });
-      }
-      const snapshot = conversationSnapshotValue(objectValue(record?.snapshot));
-      const appendedMessage = record?.appendedMessage;
-      const appendedActiveTurnId = record?.activeTurnId;
-      if (snapshot) {
-        latest = snapshot;
-        latestSequence = event.sequence;
-        for (const [turnId, sessionId] of turnProviderSessionIdsValue(record?.recovery)) {
-          turnSessions.set(turnId, sessionId);
-        }
-        if (snapshot.activeTurnId && !turnSessions.has(snapshot.activeTurnId)) {
-          const activeSession = [...sessions.values()].find((session) => session.state === "active");
-          turnSessions.set(snapshot.activeTurnId, activeSession?.id ?? null);
-        }
-      } else if (latest && isConversationMessage(appendedMessage)) {
-        if (!latest.messages.some((message) => message.id === appendedMessage.id)) {
-          latest.messages.push(structuredClone(appendedMessage));
-        }
-        if (isString(appendedActiveTurnId) || appendedActiveTurnId === null) {
-          latest.activeTurnId = appendedActiveTurnId;
-        }
-        latestSequence = event.sequence;
-      }
-    }
-    if (!latest) throw new Error(`Thread ${threadId} has no conversation events to replay.`);
-
-    db.exec("BEGIN IMMEDIATE");
-    try {
-      db.prepare("DELETE FROM projection_thread_activities WHERE thread_id = ?").run(threadId);
-      db.prepare("DELETE FROM projection_turns WHERE thread_id = ?").run(threadId);
-      db.prepare("DELETE FROM projection_thread_messages WHERE thread_id = ?").run(threadId);
-      db.prepare("DELETE FROM projection_thread_summaries WHERE thread_id = ?").run(threadId);
-      db.prepare("DELETE FROM projection_provider_sessions WHERE thread_id = ?").run(threadId);
-      db.prepare("DELETE FROM projection_attachments WHERE owner_kind = 'thread-message' AND owner_id LIKE ?").run(
-        `${threadId}:%`,
-      );
-      const sessionInsert = db.prepare(`
-        INSERT INTO projection_provider_sessions (
-          id, thread_id, provider, external_session_id, model, effort, state,
-          created_at, updated_at, resume_cursor, last_event_sequence
-        ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-      `);
-      for (const session of sessions.values()) {
-        sessionInsert.run(
-          session.id,
-          threadId,
-          session.provider,
-          session.externalSessionId,
-          session.model,
-          session.effort,
-          session.state,
-          session.createdAt,
-          session.updatedAt,
-          session.resumeCursor,
-          session.sequence,
-        );
-      }
-      const messageInsert = db.prepare(`
-        INSERT INTO projection_thread_messages (
-          thread_id, message_id, turn_id, author, status, item_type, created_at,
-          ordinal, message_json, last_event_sequence
-        ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-      `);
-      latest.messages.forEach((message, ordinal) => {
-        messageInsert.run(
-          threadId,
-          message.id,
-          message.turnId ?? null,
-          message.author,
-          message.status,
-          message.itemType ?? null,
-          message.createdAt,
-          ordinal,
-          JSON.stringify(message),
-          latestSequence,
-        );
-        for (const attachment of message.attachments ?? []) {
-          db.prepare(`
-            INSERT INTO projection_attachments
-              (attachment_id, owner_kind, owner_id, name, path, metadata_json, created_at, last_event_sequence)
-            VALUES (?, 'thread-message', ?, ?, '', ?, ?, ?)
-          `).run(
-            `${threadId}:${message.id}:${attachment.id}`,
-            `${threadId}:${message.id}`,
-            attachment.name,
-            JSON.stringify(attachment),
-            message.createdAt,
-            latestSequence,
-          );
-        }
-      });
-      const messagesByTurn = new Map<string, ConversationMessage[]>();
-      for (const message of latest.messages) {
-        if (!message.turnId) continue;
-        const messages = messagesByTurn.get(message.turnId) ?? [];
-        messages.push(message);
-        messagesByTurn.set(message.turnId, messages);
-      }
-      const turnInsert = db.prepare(`
-        INSERT INTO projection_turns (
-          turn_id, thread_id, provider_session_id, status, started_at, completed_at, last_event_sequence
-        ) VALUES (?, ?, ?, ?, ?, ?, ?)
-      `);
-      for (const [turnId, messages] of messagesByTurn) {
-        const running = latest.activeTurnId === turnId;
-        const status = running
-          ? "running"
-          : messages.some((message) => message.status === "failed")
-            ? "failed"
-            : messages.some((message) => message.status === "interrupted")
-              ? "interrupted"
-              : "completed";
-        turnInsert.run(
-          turnId,
-          threadId,
-          turnSessions.get(turnId) ?? null,
-          status,
-          messages[0]?.createdAt ?? new Date().toISOString(),
-          running ? null : (messages.at(-1)?.createdAt ?? new Date().toISOString()),
-          latestSequence,
-        );
-      }
-      const summaryInsert = db.prepare(`
-        INSERT INTO projection_thread_summaries (
-          summary_id, thread_id, through_message_id, summary_text,
-          estimated_tokens, created_at, last_event_sequence
-        ) VALUES (?, ?, ?, ?, ?, ?, ?)
-      `);
-      for (const summary of summaries) {
-        summaryInsert.run(
-          summary.id,
-          threadId,
-          summary.throughMessageId,
-          summary.text,
-          summary.estimatedTokens,
-          summary.createdAt,
-          summary.sequence,
-        );
-      }
-      const activityInsert = db.prepare(`
-        INSERT INTO projection_thread_activities (
-          activity_id, thread_id, turn_id, activity_type,
-          payload_json, created_at, last_event_sequence
-        ) VALUES (?, ?, NULL, ?, ?, ?, ?)
-      `);
-      for (const event of events) {
-        const eventPayload = JSON.parse(event.payload_json);
-        const eventRecord = objectValue(eventPayload);
-        const activityPayload =
-          conversationSnapshotValue(objectValue(eventRecord?.snapshot)) ||
-          isConversationMessage(eventRecord?.appendedMessage)
-            ? (eventRecord?.detail ?? {})
-            : eventPayload;
-        activityInsert.run(
-          randomUUID(),
-          threadId,
-          event.event_type,
-          JSON.stringify(activityPayload),
-          event.occurred_at,
-          event.sequence,
-        );
-      }
-      db.prepare(
-        `UPDATE projection_threads
-         SET active_turn_id = ?, updated_at = ?, last_event_sequence = ? WHERE thread_id = ?`,
-      ).run(latest.activeTurnId, new Date().toISOString(), latestSequence, threadId);
-      db.exec("COMMIT");
-    } catch (error) {
-      db.exec("ROLLBACK");
-      throw error;
-    }
-    return this.#conversations.readConversation(thread.agent_id, threadId);
+    return this.#replay.rebuildThreadProjection(threadId);
   }
 
   replaceMailboxState(
@@ -734,135 +302,6 @@ export class OpenBotDatabase {
   readMailboxState(): unknown | null {
     return this.#mailbox.readMailboxState();
   }
-}
-
-function providerSessionValue(value: DynamicRecord | null): ProviderSession | null {
-  if (
-    !value ||
-    !isString(value.id) ||
-    !isString(value.threadId) ||
-    !isAgentProvider(value.provider) ||
-    !isString(value.externalSessionId) ||
-    !isString(value.model) ||
-    !isString(value.effort) ||
-    (value.state !== "active" && value.state !== "inactive" && value.state !== "failed") ||
-    !isString(value.createdAt) ||
-    !isString(value.updatedAt) ||
-    (!isString(value.resumeCursor) && value.resumeCursor !== null)
-  ) {
-    return null;
-  }
-  return {
-    id: value.id,
-    threadId: value.threadId,
-    provider: value.provider,
-    externalSessionId: value.externalSessionId,
-    model: value.model,
-    effort: value.effort,
-    state: value.state,
-    createdAt: value.createdAt,
-    updatedAt: value.updatedAt,
-    resumeCursor: value.resumeCursor,
-  };
-}
-
-function summaryValue(value: DynamicRecord | null): StoredThreadSummary | null {
-  if (
-    !value ||
-    !isString(value.id) ||
-    !isString(value.threadId) ||
-    (!isString(value.throughMessageId) && value.throughMessageId !== null) ||
-    !isString(value.text) ||
-    !isNumber(value.estimatedTokens) ||
-    !isString(value.createdAt)
-  ) {
-    return null;
-  }
-  return {
-    id: value.id,
-    threadId: value.threadId,
-    throughMessageId: value.throughMessageId,
-    text: value.text,
-    estimatedTokens: value.estimatedTokens,
-    createdAt: value.createdAt,
-  };
-}
-
-function conversationSnapshotValue(value: DynamicRecord | null): ConversationSnapshot | null {
-  if (
-    !value ||
-    !isString(value.botId) ||
-    (!isString(value.threadId) && value.threadId !== null) ||
-    (!isString(value.activeTurnId) && value.activeTurnId !== null) ||
-    !isNumber(value.revision) ||
-    !Array.isArray(value.messages)
-  ) {
-    return null;
-  }
-  const messages = value.messages.filter(isConversationMessage);
-  if (messages.length !== value.messages.length) return null;
-  return {
-    botId: value.botId,
-    threadId: value.threadId,
-    activeTurnId: value.activeTurnId,
-    revision: value.revision,
-    messages,
-  };
-}
-
-function conversationRecoveryState(
-  db: DatabaseSync,
-  threadId: string,
-  activeTurnId: string | null,
-): { turnProviderSessionIds: Record<string, string | null> } {
-  const turnProviderSessionIds: Record<string, string | null> = {};
-  for (const row of databaseRows(
-    db.prepare("SELECT turn_id, provider_session_id FROM projection_turns WHERE thread_id = ?").all(threadId),
-  )) {
-    const turnId = requiredStringColumn(row, "turn_id");
-    turnProviderSessionIds[turnId] = optionalStringColumn(row, "provider_session_id");
-  }
-  if (activeTurnId && !(activeTurnId in turnProviderSessionIds)) {
-    const activeSession = databaseRow(
-      db
-        .prepare(
-          `SELECT id FROM projection_provider_sessions
-           WHERE thread_id = ? AND state = 'active'
-           ORDER BY created_at DESC, last_event_sequence DESC LIMIT 1`,
-        )
-        .get(threadId),
-    );
-    turnProviderSessionIds[activeTurnId] = activeSession ? requiredStringColumn(activeSession, "id") : null;
-  }
-  return { turnProviderSessionIds };
-}
-
-function turnProviderSessionIdsValue(value: unknown): Array<[string, string | null]> {
-  const recovery = objectValue(value);
-  const turnProviderSessionIds = objectValue(recovery?.turnProviderSessionIds);
-  if (!turnProviderSessionIds) return [];
-  const result: Array<[string, string | null]> = [];
-  for (const [turnId, sessionId] of Object.entries(turnProviderSessionIds)) {
-    if (sessionId === null || isString(sessionId)) result.push([turnId, sessionId]);
-  }
-  return result;
-}
-
-function pruneConversationSnapshots(db: DatabaseSync, threadId: string, retainedSequence: number): void {
-  db.prepare(
-    `DELETE FROM projection_thread_activities
-     WHERE thread_id = ? AND last_event_sequence IN (
-       SELECT sequence FROM orchestration_events
-       WHERE aggregate_type = 'thread' AND aggregate_id = ? AND sequence < ?
-         AND json_type(payload_json, '$.snapshot') = 'object'
-     )`,
-  ).run(threadId, threadId, retainedSequence);
-  db.prepare(
-    `DELETE FROM orchestration_events
-     WHERE aggregate_type = 'thread' AND aggregate_id = ? AND sequence < ?
-       AND json_type(payload_json, '$.snapshot') = 'object'`,
-  ).run(threadId, retainedSequence);
-  deleteOrphanReceipts(db);
 }
 
 export function providerForStoredModel(model: BotSummary["model"]): AgentProviderId {
